// src/service/aiAndProxy/schemas/lecturePlanResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Ai and Proxy API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { LecturePlanResultCreationType } from './lecturePlanResultCreationType';
import type { LecturePlanResultIndoorType } from './lecturePlanResultIndoorType';
import type { LecturePlanResultActivityTime } from './lecturePlanResultActivityTime';
import type { LecturePlanResultCourse } from './lecturePlanResultCourse';
import type { LecturePlanCardSectionResult } from './lecturePlanCardSectionResult';
import type { DriveItemResult } from './driveItemResult';
import type { SmartFolderItemResult } from './smartFolderItemResult';

/**
 * 놀이계획이면 놀이 계획 정보
 * @nullable
 */
export type LecturePlanResult = {
  /** 놀이계획의 id */
  id: number;
  /** AI의 생성 초점 타입 (TYPE_A~C) */
  creationType: LecturePlanResultCreationType;
  /** 실내외 구분(BOTH, INDOOR, OUTDOOR) */
  indoorType: LecturePlanResultIndoorType;
  /**
   * 놀이계획의 주제.
   * @maxLength 20
   */
  subject: string;
  /** 활동시간 타입 */
  activityTime: LecturePlanResultActivityTime;
  /** 활동시간 타입의 이름 */
  activityTimeStr: string;
  /** 놀이 계획의 대상 연령 */
  studentAge: number;
  /** 놀이 계획의 대상 과정 */
  course: LecturePlanResultCourse;
  /**
   * 놀이계획 시작일(yyyy-MM-dd)
   * @maxLength 10
   */
  startsAt: string;
  /**
   * 놀이계획 종료일(yyyy-MM-dd)
   * @maxLength 10
   */
  endsAt: string;
  /**
   * 주요활동의 이름들. @!,$!로 구분. 없으면 주요활동 섹션에서 따로 체크해서 넣을 예정
   * @maxLength 80
   */
  activityNames?: string;
  /**
   * 소유자 계정 아이디
   * @minimum 0
   */
  ownerAccountId: number;
  /**
   * 소유자 프로필 아이디
   * @minimum 0
   */
  ownerProfileId: number;
  /**
   * 최초 생성자 계정 아이디(공유된 것을 저장하는 게 아니라면 소유자와 동일, 공유된 것이라면 공유한 사람의 정보)
   * @minimum 0
   */
  creatorAccountId: number;
  /**
   * 최초 생성자 프로필 아이디(공유된 것을 저장하는 게 아니라면 소유자와 동일, 공유된 것이라면 공유한 사람의 정보)
   * @minimum 0
   */
  creatorProfileId: number;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  modifiedAt?: string;
  /** 놀이 계획에서 수정할 카드를 포함한 하위 섹션들. */
  lecturePlanCardSections?: LecturePlanCardSectionResult[];
  driveItem?: DriveItemResult;
  smartFolderItem?: SmartFolderItemResult;
} | null;